requirement.txt
appdirs==1.4.4
certifi==2023.11.17
charset-normalizer==3.3.2
future==0.18.3
geographiclib==2.0
idna==3.4
numpy==1.24.4
pandas==2.0.3
Pillow==9.4.0
py-staticmaps==0.4.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
requests==2.31.0
s2sphere==0.2.5
six==1.16.0
svgwrite==1.4.3
text-unidecode==1.3
tzdata==2023.3
urllib3==2.1.0
